Add JdbcCTService.
